+++ cores/domain/src/main/java/com/theloungemembers/core/api/repository/ApiAccessibleIpJpaRepository.java
@@ -5,5 +5,5 @@ import com.theloungemembers.core.api.entity.ApiAccessibleIpEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ApiAccessibleIpJpaRepository extends JpaRepository<ApiAccessibleIpEntity, Integer>, JpaSpecificationExecutor<ApiAccessibleIpEntity> {
+public interface ApiAccessibleIpJpaRepository extends JpaRepository<ApiAccessibleIpEntity, Long>, JpaSpecificationExecutor<ApiAccessibleIpEntity> {
 }
